Hand modify batch_mutate thrift code to handle errors.
I wouldn't usually modify autogenerated code, but without modifying
thrift itself this seems to be the best way to ensure errors are handled
properly for request deferreds. I've started with the batch_mutate call
as the write path seems to be the case where we generate these kinds of
errors the most.