Robustify error delivery during shutdown.
So telephus can have errors (i.e. Failures) still awaiting delivery
when it's shutting down.  Unfortunately the code it uses to do that
tries to use a member variable that's already been nulled out to
reach the reactor to asynchronously send the errback.

This adds a clause to catch that case and try to synchronously
deliver the failure to the client.